Extract MesgExtensions out of MessageWrapper. Show Save button after editing any message. DataGrid cells use OneWay binding (from data to gui, with Converter.Convert). Manually handle other direction instead of with Converter.ConvertBack.
