+++ Ui/FitEdit.Ui/Views/RecordView.axaml
@@ -67,7 +67,7 @@
 
       <Button Margin="0"
         Command="{Binding SplitActivity}"
-        IsVisible="{Binding HasSelection}"
+        IsVisible="{Binding HasNonfirstSelection}"
         Content="Split"/>
     </StackPanel>
